requirements.txt
asgiref==3.4.1
Django==4.0
django-tinymce==3.4.0
django-versatileimagefield==2.2
Pillow==8.4.0
# python-magic==0.4.24
# python-magic-bin==0.4.14
sqlparse==0.4.2
tzdata==2021.5
